backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2

# Core dependencies
numpy==1.24.3
matplotlib==3.7.2
nibabel==5.1.0
python-dotenv==1.0.0
Pillow==10.0.0
PyWavelets==1.8.0
pandas==2.3.3

# Supabase
supabase==2.0.3

# PyTorch CPU-only (much smaller, no CUDA/cuDNN)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu